gfa-extract
Removed partial fasta mapping functionality.
Both directions are fully explored (nodes on fwd do not affect nodes on bwd exploration)

SMR Optional reuse of precomputed reductions for read_from_file and process_from_memory.

+++ src/sglib/SMR.h
@@ -140,7 +140,7 @@ public:
      * @return
      * Filtered vector of RecordType elements
      */
-    std::vector<RecordType> read_from_file(const std::string &read_file) {
+    std::vector<RecordType> read_from_file(const std::string &read_file, const bool do_work=true) {
         std::string readFileBasename(read_file.substr(0,read_file.find_last_of('.')).substr(read_file.find_last_of('/')+1));
         std::string finalFilePath(outdir+readFileBasename);
         tmpInstance = tmpBase+readFileBasename;
@@ -148,7 +148,7 @@ public:
         outdir = finalFilePath;
         tmpInstance = sglib::create_temp_directory(tmpBase);
         std::ifstream final_file(finalFilePath+"final.kc");
-        if (final_file.is_open()){
+        if (final_file.is_open() and !do_work){
             sglib::OutputLog(sglib::DEBUG) << "Using precomputed sum file at " << outdir << "final.kc" << std::endl;
             return readFinalkc(outdir+"final.kc");
         } else {
@@ -179,28 +179,35 @@ public:
      * @return
      * Filtered vector of RecordType elements
      */
-    std::vector<RecordType> process_from_memory() {
+    std::vector<RecordType> process_from_memory(const bool do_work = true) {
         tmpInstance = sglib::create_temp_directory(tmpBase) + "/";
+        std::ifstream final_file(outdir+"final.kc");
+        if (final_file.is_open() and !do_work) {
+            sglib::OutputLog(sglib::DEBUG) << "Using precomputed sum file at " << outdir << "final.kc" << std::endl;
+            return readFinalkc(outdir + "final.kc");
+        } else {
+            uint64_t numFileRecords(0);
 
-        uint64_t numFileRecords(0);
+            std::chrono::time_point<std::chrono::system_clock> start, end;
+            start = std::chrono::system_clock::now();
 
-        std::chrono::time_point<std::chrono::system_clock> start, end;
-        start = std::chrono::system_clock::now();
+            sglib::OutputLog(sglib::DEBUG) << "Reading From memory" << std::endl;
+            FileReader myFileReader(reader_parameters);
+            sglib::OutputLog(sglib::DEBUG) << "Begin reduction using " << numElementsPerBatch << " elements per batch ("
+                                           << ceil(uint64_t((numElementsPerBatch * sizeof(RecordType) * maxThreads)) /
+                                                   (1.0f * 1024 * 1024 * 1024)) << "GB)" << std::endl;
+            mapElementsToBatches(myFileReader, numFileRecords);
 
-        sglib::OutputLog(sglib::DEBUG) << "Reading From memory"<< std::endl;
-        FileReader myFileReader(reader_parameters);
-        sglib::OutputLog(sglib::DEBUG) << "Begin reduction using " << numElementsPerBatch << " elements per batch (" << ceil(uint64_t((numElementsPerBatch*sizeof(RecordType)*maxThreads)) / (1.0f*1024*1024*1024)) << "GB)" << std::endl;
-        mapElementsToBatches(myFileReader, numFileRecords);
+            readerStatistics = myFileReader.getSummaryStatistics();
 
-        readerStatistics = myFileReader.getSummaryStatistics();
-
-        end = std::chrono::system_clock::now();
-        std::chrono::duration<double> elapsed_seconds = end - start;
-        sglib::OutputLog(sglib::DEBUG) << "Done reduction in " << elapsed_seconds.count() << "s" << std::endl;
-        //TODO: remove instance / never create the final files?
-        std::vector<RecordType> result(getRecords());
-        sglib::remove_directory(tmpInstance);
-        return result;
+            end = std::chrono::system_clock::now();
+            std::chrono::duration<double> elapsed_seconds = end - start;
+            sglib::OutputLog(sglib::DEBUG) << "Done reduction in " << elapsed_seconds.count() << "s" << std::endl;
+            //TODO: remove instance / never create the final files?
+            std::vector<RecordType> result(getRecords());
+            sglib::remove_directory(tmpInstance);
+            return result;
+        }
     };
 
     /**
